+++ src/main/java/artisanSpringBoot/model/MetierJoinCompte.java
@@ -16,7 +16,7 @@ public class MetierJoinCompte {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE,generator = "seqImplementation")
-	private Long id;
+	private Integer id;
 	@ManyToOne
 	@JoinColumn(name="id_compte", foreignKey = @ForeignKey(name="metier_compte_id_fk"))
 	private Compte compte;
@@ -29,12 +29,12 @@ public class MetierJoinCompte {
 		
 	}
 
-	public Long getId() 
+	public Integer getId() 
 	{
 		return id;
 	}
 
-	public void setId(Long id) 
+	public void setId(Integer id) 
 	{
 		this.id = id;
 	}
